Move expense out of movement
