+++ citywok_ms/templates/expense/order_payment.html
@@ -29,7 +29,7 @@
                 {% endwith %}
             </div>
             {% with money_form=form.value %}
-            {% include 'movement/expense/money.html' %}
+            {% include 'expense/money.html' %}
             {% endwith %}
             {% if request.endpoint == 'expense.new_order_payment' %}
             <div>
